version bump and refactor for packaging
